Changes to handle cases where a user logs in over multiple devices
Until now, all updates to a user's view were driven by the user's
actions (such as adding users or adding projects). When we update the
project state externally (such as through a github hook), we need to
post that change to the user's page. This was the right time to
therefore also add support for multiple sign-ins. We maintain an active
set of sockets for a user and notify across all those sockets when a
project update happens. I need to extend this  to all actions, so that
when a new project is added, I emit the update on all of that user's
currently active sockets.